new stuff to plugin.py
