+++ xiplot/plugin.py
@@ -1,15 +1,11 @@
 """
-    This module exposes the parts of xiplot that are useful for implementing
-    plugins. Most of the members of this module are just imported from other
-    places in xiplot. But there are also some plugin-specific functions and
-    objects.
+    This module exposes the parts of xiplot that are useful for implementing plugins.
+    Most of the members of this module are just imported from other places in xiplot.
 """
 
-from importlib.metadata import entry_points as _entry_points
 from io import BytesIO
 from os import PathLike
-from typing import Any, Callable, Dict, List, Literal, Tuple, Union
-from warnings import warn
+from typing import Any, Callable, Dict, Tuple, Union
 
 import pandas as pd
 from dash import Dash
@@ -18,7 +14,16 @@ from dash.development.base_component import Component
 # Export useful parts of xiplot:
 from xiplot.plots import APlot  # noqa: F401
 from xiplot.utils import generate_id  # noqa: F401
+from xiplot.utils.auxiliary import (  # noqa: F401
+    decode_aux,
+    encode_aux,
+    get_clusters,
+    get_selected,
+    merge_df_aux,
+)
+from xiplot.utils.cluster import cluster_colours  # noqa: F401
 from xiplot.utils.components import (  # noqa: F401
+    ClusterDropdown,
     ColumnDropdown,
     DeleteButton,
     FlexRow,
@@ -35,8 +40,7 @@ ID_HOVERED = "lastly_hovered_point_store"
 ID_CLICKED = "lastly_clicked_point_store"
 # `dcc.Store` that is `True` if the dark mode is active
 ID_DARK_MODE = "light-dark-toggle-store"
-# `dcc.Store` that contains the current plotly template
-#  (used for the dark mode)
+# `dcc.Store` that contains the current plotly template (used for the dark mode)
 ID_PLOTLY_TEMPLATE = "plotly-template"
 
 # Useful CSS classes
@@ -86,107 +90,3 @@ def placeholder_figure(text: str) -> Dict[str, Any]:
             ],
         }
     }
-
-
-def get_plugins(
-    plugin_type: Literal["read", "write", "plot", "global", "callback"],
-) -> List[Tuple[str, str, Any]]:
-    """Get a list of all plugins of the specified type.
-
-    Args:
-        plugin_type: The type of xiplot plugin.
-
-    Returns:
-        A list of tuples of each plugin's name, path, and its
-        loaded `entry_point`.
-    """
-    try:
-        # Python 3.10+
-        plugins = _entry_points(group=f"xiplot.plugin.{plugin_type}")
-    except TypeError:
-        # Python 3.8-3.9
-        plugins = _entry_points().get(f"xiplot.plugin.{plugin_type}", ())
-
-    loaded_plugins = []
-    for plugin in plugins:
-        try:
-            loaded_plugins.append((plugin.name, plugin.value, plugin.load()))
-        except Exception as e:
-            warn(f"Could not load plugin {plugin}: {e}")
-
-    return loaded_plugins
-
-
-def get_plugins_cached(
-    plugin_type: Literal["read", "write", "plot", "global", "callback"],
-) -> List[Tuple[str, str, Any]]:
-    """Get a list of all plugins of the specified type
-     (this call is cached for future reuse).
-
-    Args:
-        plugin_type: The type of xiplot plugin.
-
-    Returns:
-        A list of tuples of each plugin's name, path, and its
-        loaded `entry_point`.
-    """
-    if getattr(get_plugins_cached, "cache", None) is None:
-        get_plugins_cached.cache = dict()
-
-    cached_plugins_of_type = get_plugins_cached.cache.get(plugin_type, None)
-
-    if cached_plugins_of_type is not None:
-        return cached_plugins_of_type
-
-    loaded_plugins = get_plugins(plugin_type)
-    get_plugins_cached.cache[plugin_type] = loaded_plugins
-
-    return loaded_plugins
-
-
-def get_all_loaded_plugins() -> List[Tuple[str, str, str, Any]]:
-    """Collects the list of all loaded plugins and their types.
-
-    Returns:
-        A list of tuples of each plugin's type, name, path, and its
-        loaded `entry_point`.
-
-    """
-    plugins = []
-
-    for plugin_type in ["read", "write", "plot", "global", "callback"]:
-        for name, path, plugin in get_plugins(plugin_type):
-            plugins.append((plugin_type, name, path, plugin))
-
-    return plugins
-
-
-def get_all_loaded_plugins_cached() -> List[Tuple[str, str, str, Any]]:
-    """Collects the list of all cached loaded plugins and their types.
-
-    Returns:
-        A list of tuples of each plugin's type, name, path, and its
-        loaded `entry_point`.
-
-    """
-    plugins = []
-
-    for plugin_type in ["read", "write", "plot", "global", "callback"]:
-        for name, path, plugin in get_plugins_cached(plugin_type):
-            plugins.append((plugin_type, name, path, plugin))
-
-    return plugins
-
-
-def is_dynamic_plugin_loading_supported() -> bool:
-    """Checks if xiplot supports dynamically loading new plugins at runtime.
-
-    Returns:
-        `True` if dynamic loading is supported, `False` otherwise.
-    """
-    try:
-        import micropip  # noqa: F401
-    except ImportError:
-        return False
-
-    return True
